Refactor Player. Now to gain flexibility we are going to base on hexagonal arquitechture.
