+++ src/MainScene.js
@@ -1,6 +1,6 @@
 import Phaser from 'phaser';
-import Player from './Player';
-import spriteSheet from '../assets/sprites/player.png';
+import { createPlayer } from './Player/Player';
+//import spriteSheet from '../assets/sprites/player.png';
 
 export default class MainScene extends Phaser.Scene{
 
@@ -22,28 +22,9 @@ export default class MainScene extends Phaser.Scene{
     }
       
     create () {
-        this.player = new Player({
-            scene: this,
-            key: 'player',
-            x: 16,
-            y: 16
-        });
-
-        /*this.load.spritesheet(
-            'playerSpritesheet', 
-            spriteSheet,
-            {
-                frameWidth: 16,
-                frameHeight: 24
-            }
-        );*/        
-        this.anims.create({
-            key: 'idle',
-            frames: this.anims.generateFrameNames('playerSpritesheet',{start: 0, end: 1}),
-            frameRate: 10,
-            repeat: -1
-        });
-
+        // populate this.player
+        createPlayer(this);
+        
         this.player.play("idle");
